requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
sqlalchemy==1.4.38
databases==0.5.4
aiomysql==0.0.22
pydantic==1.9.1
bcrypt==3.2.0
jose==3.3.0
python-dotenv==1.0.0
openai>=1.0.0
